Move withdraw brief view to app.main.views.withdraw_brief
